Update on "[dynamo] Move descriptor VTs to descriptors.py"
Move all descriptor-related VariableTracker classes from misc.py to a
new dedicated file _dynamo/variables/descriptors.py. This includes:

  Non-data descriptors:
  - WrapperDescriptorVariable, MethodWrapperVariable
  - MethodDescriptorVariable, BuiltinMethodVariable
  - ClassMethodDescriptorVariable
  - StaticMethodVariable, ClassMethodVariable

  Data descriptors:
  - MemberDescriptorVariable
  - GetSetDescriptorVariable
  - PropertyVariable

No behavioral changes -- pure code movement.

Authored with Claude.

[ghstack-poisoned]
