+++ torch/_dynamo/variables/__init__.py
@@ -62,26 +62,36 @@ from .dicts import (
 from .distributed import BackwardHookVariable, DistributedVariable
 from .functions import (
     BaseUserFunctionVariable,
+    BoundBuiltinMethodVariable,
     BuiltinMethodVariable,
+    ClassMethodDescriptorVariable,
+    ClassMethodVariable,
     CollectionsNamedTupleFunction,
     CreateTMADescriptorExperimentalVariable,
     CreateTMADescriptorStableVariable,
     FunctionDecoratedByContextlibContextManagerVariable,
     FunctoolsPartialVariable,
+    GetSetDescriptorVariable,
     InspectSignatureVariable,
     LocalGeneratorFunctionVariable,
     LocalGeneratorObjectVariable,
+    MemberDescriptorVariable,
+    MethodDescriptorVariable,
+    MethodWrapperVariable,
     NestedUserFunctionVariable,
     PolyfilledFunctionVariable,
+    PropertyVariable,
     PyTreeGetNodeTypeFunctionVariable,
     PyTreeTreeIsLeafFunctionVariable,
     SkipFunctionVariable,
     SparseTensorCreationSkipVariable,
+    StaticMethodVariable,
     TMADescriptorExperimentalVariable,
     TMADescriptorStableVariable,
     TritonSetAllocatorVariable,
     UserFunctionVariable,
     UserMethodVariable,
+    WrapperDescriptorVariable,
     WrapperUserFunctionVariable,
     WrapperUserMethodVariable,
 )
@@ -109,18 +119,6 @@ from .lists import (
     SliceVariable,
     TupleIteratorVariable,
     TupleVariable,
-)
-from .descriptors import (
-    BuiltinMethodVariable,
-    ClassMethodDescriptorVariable,
-    ClassMethodVariable,
-    GetSetDescriptorVariable,
-    MemberDescriptorVariable,
-    MethodDescriptorVariable,
-    MethodWrapperVariable,
-    PropertyVariable,
-    StaticMethodVariable,
-    WrapperDescriptorVariable,
 )
 from .misc import (
     AutogradFunctionContextVariable,
